refactor(labkit): 把确定性内核与机器层上移成两个实验台共用的地基
gpulab（见 design/gpulab.md）要复用 opslab 已经建好的 L0 与 L1：虚拟时钟、
优先级定序、settle 与死锁检测、种子 RNG、VFS、tree-sitter shell 与 coreutils。
不动的话 gpulab 得写 `import { Kernel } from '../opslab/kernel'` —— 一个 GPU
项目依赖 K8s 项目，读起来就是错的。

这一版只做移动，不做抽象：

  src/lib/opslab/kernel/          → src/lib/labkit/kernel/        （630 行，整体）
  src/lib/opslab/machine/vfs.ts   → src/lib/labkit/machine/vfs.ts
  src/lib/opslab/machine/shell/   → src/lib/labkit/machine/shell/
  src/lib/opslab/machine/machine.ts → src/lib/labkit/machine/machine.ts
  src/lib/opslab/machine/oci/     → src/lib/opslab/oci/           （k8s 专用，留下）
  tests/opslab/{kernel,machine}.test.ts → tests/labkit/

之所以不趁机设计一层「实验台」抽象：两个消费者是好抽象的最低样本量，但你是在
建第二个消费者的过程中才知道该抽什么。现在设计接口等于在没有第二个用例的情况下
猜接缝，猜错比不抽还贵。真正的抽象留到 gpulab 跑起来之后再提取。

oci 一并从 machine/ 里提出来：留在原地的话，`machine/` 这个目录里就只剩镜像构建
和 registry，没有机器 —— 那比现在更难读。它依赖 labkit 的 vfs 与 shell，方向正确。

被移动的文件本身一行没改（clock/kernel/random/machine/vfs/shell 的 diff 都是 0），
改的只有两个 barrel 的注释、以及 49 个文件里的 import 路径。

顺手修掉一个会静默失效的地方：scripts/test-runner.js 里的套件是按目录写死的
（'tests/opslab'），测试文件移到 tests/labkit 之后 npm test 会照样全绿，
但那 87 个用例一个都没跑。现在拆成 Labkit Foundation 与 Opslab Cluster 两个套件。

验证：tsc 干净、npm test 9/9 套件、npm run projects:verify 全绿、next build 通过。
