+++ src/lib/labkit/machine/shell/parser.ts
