+++ src/lib/labkit/machine/shell/shell.ts
