+++ src/lib/labkit/machine/vfs.ts
