+++ tests/opslab/cli-runtime.test.ts
@@ -11,7 +11,7 @@ import fs from 'node:fs';
 import path from 'node:path';
 import { createRequire } from 'node:module';
 import { createCluster } from '../../src/lib/opslab/controllers';
-import { createMachine, Machine } from '../../src/lib/opslab/machine';
+import { createMachine, Machine } from '../../src/lib/labkit/machine';
 import {
   CliRuntime, createCliRuntime, createIndexedDbCache, installClusterCli,
   renderKubeconfig, defaultKubeconfig, remoteSignature,
